feat(app): Add FlexRadioGroup Library (forgot to add build.gradle in last commit). Set manually radioButtons since is not working in XML. Remove 2 tests from SoundLibraryFragmentTest.kt and add 1 test to RandomPresetFragmentTest.kt
