+++ app/build.gradle
@@ -83,13 +83,14 @@ dependencies {
   implementation 'com.github.medyo:android-about-page:1.2.5'
   implementation 'com.google.android.exoplayer:exoplayer-core:2.11.8'
   implementation 'com.google.android:flexbox:2.0.1'
+  implementation 'com.hopenlib.library:flextools:1.0.1'
   implementation 'com.google.android.material:material:1.3.0'
   implementation 'com.google.code.gson:gson:2.8.6'
   implementation 'io.noties.markwon:core:4.6.2'
   implementation 'org.greenrobot:eventbus:3.1.1'
   implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
 
-  playstoreImplementation 'com.anjlab.android.iab.v3:library:1.1.0'
+    playstoreImplementation 'com.anjlab.android.iab.v3:library:1.1.0'
   playstoreImplementation 'com.google.android.gms:play-services-cast-framework:19.0.0'
   playstoreImplementation 'com.google.android.play:core:1.10.0'
